Move icon/diagnostics/ to diagnostics/icon/
Update all absolute imports from jcm.physics.icon.diagnostics to
jcm.physics.diagnostics.icon.
