+++ jcm/physics/diagnostics/icon/__init__.py
@@ -4,7 +4,7 @@ This module contains diagnostic calculations and utility functions
 used throughout the ICON physics parameterizations.
 """
 
-from jcm.physics.icon.diagnostics.wmo_tropopause import wmo_tropopause
+from jcm.physics.diagnostics.icon.wmo_tropopause import wmo_tropopause
 
 __all__ = [
     'wmo_tropopause',
